chore: separates segment and subsegment and includes facade segment.
