+++ tests/Unit/SegmentTest.php
@@ -3,13 +3,12 @@
 namespace AwsXRayTests\Unit;
 
 use AwsXRay\Header;
-use AwsXRay\Plugins\PluginMetadata;
-use AwsXRay\Segment;
-use InvalidArgumentException;
-use PHPUnit\Framework;
 use AwsXRay\Recorder;
+use PHPUnit\Framework;
+use AwsXRay\ParentSegment;
+use InvalidArgumentException;
 
-final class SegmentTest extends Framework\TestCase
+final class ParentSegmentTest extends Framework\TestCase
 {
     const NAME = 'test';
     const CHILD_NAME = 'test_child';
@@ -18,7 +17,7 @@ final class SegmentTest extends Framework\TestCase
 
     public function testCreateSegmentWithoutHeader()
     {
-        $segment = Segment::create(new Recorder(), self::NAME);
+        $segment = ParentSegment::create(new Recorder(), self::NAME);
 
         $this->assertEquals(16, strlen($segment->getId()));
         $this->assertEquals(1 + 1 + 8 + 1 + 24, strlen($segment->getTraceId()));
@@ -36,7 +35,7 @@ final class SegmentTest extends Framework\TestCase
     {
         $header = new Header(self::TRACE_ID, self::PARENT_ID);
 
-        $segment = Segment::create(new Recorder(), self::NAME, $header);
+        $segment = ParentSegment::create(new Recorder(), self::NAME, $header);
 
         $this->assertArraySubset(
             [
@@ -51,50 +50,16 @@ final class SegmentTest extends Framework\TestCase
 
     public function testAddsAnnotationsSuccess()
     {
-        $segment = Segment::create(new Recorder(), self::NAME);
+        $segment = ParentSegment::create(new Recorder(), self::NAME);
         $segment->addAnnotation('key', 'value');
         $this->assertEquals(['key' => 'value'], $segment->jsonSerialize()['annotations']);
     }
 
-    public function testAddsAnnotationsFails()
-    {
-        $this->expectException(InvalidArgumentException::class);
-        $segment = Segment::create(new Recorder(), self::NAME);
-        $segment->addAnnotation('key', ['a']);
-    }
-
     public function testCloseSuccess()
     {
-        $segment = Segment::create(new Recorder(), self::NAME);
+        $segment = ParentSegment::create(new Recorder(), self::NAME);
         $segment->close();
         $this->assertArrayHasKey('end_time', $segment->jsonSerialize());
         $this->assertArrayNotHasKey('in_progress', $segment->jsonSerialize());
-    }
-
-    public function testAddPlugin()
-    {
-        $metadata = [
-            'elastic_beanstalk' => [
-                'environment_name' => 'test_environment_name',
-                'version_label' => 'test_version_label',
-                'deployment_id' => 'test_deployment_id',
-            ],
-            'ec2' => [
-                'instance_id' => 'test_instance_id',
-                'availability_zone' => 'test_availability_zone',
-            ],
-            'ecs' => [
-                'container' => 'test_container',
-            ],
-            'origin' => 'test_origin',
-        ];
-        $metadataPlugin = PluginMetadata::create($metadata);
-
-        $segment = Segment::create(new Recorder(), 'test_name');
-        $segment->addPlugin($metadataPlugin);
-        $this->assertEquals('test_origin', $segment->jsonSerialize()['origin']);
-        $this->assertEquals($metadata['elastic_beanstalk'], $segment->jsonSerialize()['aws']['elastic_beanstalk']);
-        $this->assertEquals($metadata['ec2'], $segment->jsonSerialize()['aws']['ec2']);
-        $this->assertEquals($metadata['ecs'], $segment->jsonSerialize()['aws']['ecs']);
     }
 }
